@@ -1,71 +1,86 @@
 {
   "name": "@iebh/methods-wizard",
-  "version": "0.5.0",
-  "private": true,
+  "description": "The Evidence Research Accelerator",
+  "version": "2.0.0",
+  "type": "module",
   "scripts": {
-    "serve": "cross-env NODE_OPTIONS=--openssl-legacy-provider vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview",
+    "start": "vite"
   },
-  "author": [
-    "Connor Forbes <cforbes.software@gmail.com> (https://github.com/connorf25)",
-    "Justin Cark <jclark@bond.edu.au> (https://github.com/Justin-Clarc)",
-    "Matt Carter <m@ttcarter.com> (https://github.com/hash-bang)"
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/IEBH/methods-wizard.git"
+  },
+  "author": {
+    "name": "Matt Carter",
+    "email": "m@ttcarter.com",
+    "url": "https://github.com/hash-bang"
+  },
+  "contributors": [
+    {
+      "name": "Connor Forbes",
+      "email": "cforbes.software@gmail.com",
+      "url": "https://github.com/connorf25"
+    },
+    {
+      "name": "Justin Cark",
+      "email": "jclark@bond.edu.au",
+      "url": "https://github.com/Justin-Clarc"
+    },
+    {
+      "name": "Tian Liang",
+      "email": "amethystoct2012@gmail.com",
+      "url": "https://github.com/Octian"
+    }
   ],
-  "license": "MIT",
-  "dependencies": {
-    "@deepstream/client": "^5.2.4",
-    "@fortawesome/fontawesome-free": "^5.14.0",
-    "core-js": "^3.6.5",
-    "element-ui": "^2.15.12",
-    "firebase": "^7.16.0",
-    "lodash": "^4.17.21",
-    "primeflex": "^2.0.0-rc.1",
-    "primeicons": "^5.0.0",
-    "primevue": "^2.9.2-rc.3",
-    "quill": "^1.3.7",
-    "vue": "^2.6.11",
-    "vue-router": "^3.3.4",
-    "vue-sidebar-menu": "^4.6.2",
-    "vue2-ace-editor": "^0.0.15",
-    "vuex": "^3.5.1",
-    "xlsx-js-style": "^1.2.0"
+  "engineStrict": true,
+  "engines": {
+    "node": ">=20"
   },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.4.0",
-    "@vue/cli-plugin-eslint": "~4.4.0",
-    "@vue/cli-service": "~4.4.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "babel-eslint": "^10.1.0",
-    "cross-env": "^7.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-vue": "^6.2.2",
-    "prettier": "^1.19.1",
-    "vue-template-compiler": "^2.6.11",
+  "private": true,
+  "license": "PRIVATE",
+  "dependencies": {
+    "@deepstream/client": "^7.0.2",
+    "@fortawesome/fontawesome-free": "^6.4.2",
+    "@iebh/tera-fy": "^1.0.4",
+    "@momsfriendlydevco/eslint-config": "^1.0.7",
+    "@vitejs/plugin-vue": "^4.4.0",
+    "bootstrap": "^5.3.2",
+    "brace": "^0.11.1",
+    "eslint": "^8.52.0",
+    "lodash-es": "^4.17.21",
+    "primeflex": "^3.3.1",
+    "primeicons": "^6.0.1",
+    "primevue": "^3.38.1",
+    "rollup-plugin-visualizer": "^5.9.2",
+    "sass": "^1.69.5",
+    "unplugin-cloudflare-headers": "^0.3.3",
+    "vite": "^4.5.0",
+    "vue": "^3.3.7",
+    "vue-eslint-parser": "^9.3.2",
+    "vue-router": "^4.2.5",
+    "vue-sidebar-menu": "^5.2.11",
+    "vue2-ace-editor": "^0.0.15",
+    "vuex": "^4.1.0",
     "xlsx": "^0.18.5"
   },
-  "prettier": {
-    "useTabs": true
-  },
   "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
+    "extends": "@momsfriendlydevco",
+    "globals": {
+      "globalThis": "writable"
     },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended",
-      "@vue/prettier"
+    "ignorePatterns": [
+      ".*",
+      "docs/**/*",
+      "dist/**/*",
+      "node_modules/**/*",
+      "scripts/**/*"
     ],
     "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+      "ecmaVersion": "latest"
+    }
+  }
 }
